Use fp32 for sin/cos in SDXL.

+++ models/experimental/stable_diffusion_xl_base/tt/tt_timesteps.py
@@ -24,10 +24,9 @@ class TtTimesteps(LightweightModule):
         exponent = -log(self.max_period) * torch.arange(start=0, end=self.half_dim, dtype=torch.float32)
         exponent = exponent / (self.half_dim - downscale_freq_shift)
 
-        # Setting emb to bfloat16 increases image quality
         self.emb = ttnn.from_torch(
             torch.exp(exponent),
-            dtype=ttnn.DataType.BFLOAT16,
+            dtype=ttnn.DataType.FLOAT32,
             layout=ttnn.TILE_LAYOUT,
             device=device,
             memory_config=ttnn.DRAM_MEMORY_CONFIG,
